[sdk33] Upgrade all SoLoader references to 0.6.0
SoLoader was upgraded to 0.6.0 in RN 0.59. It is backwards-compatible with SoLoader 0.5.1, which was not backwards-compatible with SoLoader 0.1.0.

This is OK, since RN 0.57 upgraded to SoLoader 0.5.1 and that is the oldest RN version we are including in the next Expo Client with SDK 31 & 32. We are dropping SDK 30 and older, which used RN 0.55.

So, we don't need to patch SoLoader 0.6.1 with hand-written JVM bytecode (see https://github.com/expo/expo/commit/7b5d52ed77fed9d6bf7c0d55ba0608b4341787ec for the patch) and include it in our local Maven repository and can just use the regular copy from Maven Central. Also, this means we can use a newer javac, which disallows .class files with two methods that differ only by return type -- the technique we were relying on to simultaneously support SoLoader 0.1.0 and 0.5.1 previously.